requirements.txt
pytorch ~= 1.4.0
numpy ~= 1.20.2
scipy ~= 1.6.2
pandas ~= 1.2.4
transformers == 3.0.2
dotenv
tqdm